+++ apps/web/src/handler/canvasCommand/resolvers/index.ts
@@ -2,6 +2,7 @@ export { default as resolveAddNodes } from './resolveAddNodes';
 export { default as resolveDisconnectEdge } from './resolveDisconnectEdge';
 export { default as resolveGroupRectIntoFrame } from './resolveGroupRectIntoFrame';
 export { default as resolveGroupSelectionIntoFrame } from './resolveGroupSelectionIntoFrame';
+export { default as resolveMoveNoteExcerpt } from './resolveMoveNoteExcerpt';
 export { default as resolveNodeDragStop } from './resolveNodeDragStop';
 export { default as resolvePasteClipboard } from './resolvePasteClipboard';
 export { default as resolveSelectNodes } from './resolveSelectNodes';